refactor(sponsor): file names and restructure project layout
