+++ content/zh_tw/sponsor/index.md
@@ -1,21 +1,5 @@
----
-title: 贊助
----
+<script setup lang="ts">
+import Page from './index.vue'
+</script>
 
-# COSCUP Taiwan 2025 贊助 {style="text-align:center"}
-
-## 贊助我們
-
-請聯絡 sponsorship@coscup.org
-
-## 第一種
-
-<Sponsor1 />
-
-## 第二種
-
-<Sponsor2 />
-
-## Footer
-
-<SponsorFooter />
+<Page />
