+++ content/zh_tw/sponsor/index.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import SponsorOneColumn from './SponsorOneColumn.vue'
+import SponsorTwoColumns from './SponsorTwoColumns.vue'
+</script>
+
+<template>
+  <SponsorOneColumn />
+  <SponsorTwoColumns />
+</template>
